fix: rate score logic

+++ internal/dto/rating.go
@@ -3,14 +3,14 @@ package dto
 import "github.com/gift-redemption/internal/model"
 
 type RatingRequest struct {
-	Score int `json:"score" binding:"required,min=1,max=5"`
+	Score float64 `json:"score" binding:"required,min=1,max=5"`
 }
 
 type RatingResponse struct {
 	ID         uint    `json:"id"`
 	GiftID     uint    `json:"gift_id"`
 	GiftName   string  `json:"gift_name"`
-	Score      int     `json:"score"`
+	Score      float64 `json:"score"`
 	AvgRating  float64 `json:"avg_rating"`
 	StarRating float64 `json:"star_rating"`
 }
